chore: Move citrea mock to test/ and make citrea module a single file.
