+++ core/src/test/common/mod.rs
@@ -13,7 +13,6 @@ use crate::builder::transaction::{
     create_replacement_deposit_txhandler, BaseDepositData, DepositInfo, DepositType,
     ReplacementDepositData, SecurityCouncil, TxHandler, DEFAULT_SEQUENCE,
 };
-use crate::citrea::mock::MockCitreaClient;
 use crate::citrea::CitreaClientT;
 use crate::config::BridgeConfig;
 use crate::errors::BridgeError;
@@ -35,7 +34,7 @@ use bitcoin::secp256k1::PublicKey;
 use bitcoin::XOnlyPublicKey;
 use bitcoin::{taproot, Amount, BlockHash, OutPoint, Transaction, Txid, Witness};
 use bitcoincore_rpc::RpcApi;
-use citrea::get_transaction_params;
+use citrea::{get_transaction_params, MockCitreaClient};
 use eyre::Context;
 use secp256k1::rand;
 pub use setup_utils::*;
